dom/nodes/Element-class-alias.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>class alias for className</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h1>Tests if the class and className IDL attributes do the same</h1>
<div id="testDiv"></div>

<script>
test(function() {
  assert_equals(testDiv.class, "", "class must reflect \"class\"");
  assert_equals(testDiv.className, "", "className must reflect \"class\"");
  
  testDiv.setAttribute("class", "test 1");
  assert_equals(testDiv.class, "test 1", "class must reflect changes made to \"class\"");
  assert_equals(testDiv.className, "test 1", "className must reflect changes made to \"class\"");
});
test(function() {
  testDiv.className = "test 2";
  assert_equals(testDiv.class, "test 2", "setting className must be reflected by class");
  assert_equals(testDiv.className, "test 2", "setting className must be reflected by className");
  assert_equals(testDiv.getAttribute("class"), "test 2", "setting className must be reflected by \"class\"");
});
test(function() {
  testDiv.class = "test 3";
  assert_equals(testDiv.class, "test 3", "setting class must be reflected by class");
  assert_equals(testDiv.className, "test 3", "setting class must be reflected by className");
  assert_equals(testDiv.getAttribute("class"), "test 3", "setting class must be reflected by \"class\"");
});
</script>
